♻️ `ActivityLifecycleCallbacksAdapter` 추가.
`ActivityLifecycleCallbacksAdapter` 추가.
