+++ app/src/main/java/com/github/hemoptysisheart/parking/app/support/ActivityLifecycleCallbacksAdapter.kt
@@ -1,0 +1,21 @@
+package com.github.hemoptysisheart.parking.app.support
+
+import android.app.Activity
+import android.app.Application
+import android.os.Bundle
+
+open class ActivityLifecycleCallbacksAdapter : Application.ActivityLifecycleCallbacks {
+    override fun onActivityCreated(activity: Activity, savedInstanceState: Bundle?) {}
+
+    override fun onActivityStarted(activity: Activity) {}
+
+    override fun onActivityResumed(activity: Activity) {}
+
+    override fun onActivityPaused(activity: Activity) {}
+
+    override fun onActivityStopped(activity: Activity) {}
+
+    override fun onActivitySaveInstanceState(activity: Activity, outState: Bundle) {}
+
+    override fun onActivityDestroyed(activity: Activity) {}
+}
